FComponent keeps track of PlatformInterface
Each component now keeps track of the PlatformInterface object used when
it was constructed.  This way we are sure that when the object is
destroyed, any platform-dependent objects are returned to the same
PlatformInterface object they came from.  Additionally, FComponent and
FUI now have constructors with PlatformInterface parameters in case you
want to create an object using a PlatformInterface other than the one
which would otherwise be returned by Platform::get().